more matching, rename Audio_SetBGM to Audio_QueueSeqCmd

+++ include/variables.h
@@ -724,7 +724,7 @@ extern u8 D_801333CC;
 extern u16 D_801333D0;
 extern Vec3f D_801333D4;
 extern f32 D_801333E0;
-extern f32 D_801333E8;
+extern s8 D_801333E8;
 extern u8 D_801333F0;
 extern u8 D_801333F4;
 extern u8 D_801333F8;
